+++ src/main/java/org/bytedeco/javacpp/PointerPointer.java
@@ -113,7 +113,7 @@ public class PointerPointer<P extends Pointer> extends Pointer {
      *
      * @param size the number of {@code void*} elements to allocate
      */
-    public PointerPointer(int size) {
+    public PointerPointer(long size) {
         try {
             allocateArray(size);
         } catch (UnsatisfiedLinkError e) {
@@ -124,7 +124,7 @@ public class PointerPointer<P extends Pointer> extends Pointer {
     public PointerPointer() { }
     /** @see Pointer#Pointer(Pointer) */
     public PointerPointer(Pointer p) { super(p); }
-    private native void allocateArray(int size);
+    private native void allocateArray(long size);
 
     /** This is just to keep references to Pointer objects and prevent premature deallocation. */
     private P[] pointerArray;
